Move test to `block_announce_test.go`
